requirements.txt
elasticsearch==7.12.0
gensim==4.3.0
jieba==0.42.1
numpy==1.23.4
pandas==1.5.2
python_Levenshtein==0.20.9
sentence_transformers==2.2.2
tornado==6.2
